+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.4.2
 * fix (notifier): error when running async (issue #2)
 * fix (backend): logging error when running async
+* fix (backend): alert processing failed if deduplication contained a '%' (issue #1)
 
 # 2.4.1
 * make reason value renderable