+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2.4.2
+* fix (notifier): error when running async (issue #2)
+* fix (backend): logging error when running async
+* fix (backend): alert processing failed if deduplication contained a '%' (issue #1)
+
 # 2.4.1
 * make reason value renderable
 * support to provide auto close/resolve alert reason with autoCloseText/autoResolve attributes